web/requirements.txt
click==6.7
Flask==1.0
flask-restful==0.3.8
itsdangerous==0.24
Jinja2==2.10
#MarkupSafe==1.0
redis==2.10.6
Werkzeug==0.15.3
wheel==0.24.0
flask_bootstrap
flask-debug
flask-reqarg
requests
mecab-python3